refactor: abstract social registration
